+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Automatically connect with simulator when seeking position
 - Also pause the initially loaded flight
 - When "repeat values" is enabled then also repeat values with "zero value" (e.g. "canopy closed")
+- Automatically reconnect in case the SimConnect server (flight simulator) has crashed
 
 ## 0.6.0